requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
albucore==0.0.12
    # via albumentations
albumentations==1.4.11
annotated-types==0.7.0
    # via pydantic
eval-type-backport==0.2.0
    # via albumentations
filelock==3.15.4
    # via
    #   torch
    #   triton
fsspec==2024.6.1
    # via torch
imageio==2.34.2
    # via scikit-image
jinja2==3.1.4
    # via torch
joblib==1.4.2
    # via scikit-learn
lazy-loader==0.4
    # via scikit-image
markupsafe==2.1.5
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via
    #   scikit-image
    #   torch
numpy==1.26.4
    # via
    #   albucore
    #   albumentations
    #   imageio
    #   opencv-python-headless
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
    #   torchvision
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.5.82
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
opencv-python-headless==4.10.0.84
    # via
    #   albucore
    #   albumentations
packaging==24.1
    # via
    #   lazy-loader
    #   scikit-image
pillow==10.4.0
    # via
    #   imageio
    #   scikit-image
    #   torchvision
pydantic==2.8.2
    # via albumentations
pydantic-core==2.20.1
    # via pydantic
pyyaml==6.0.1
    # via albumentations
scikit-image==0.24.0
    # via albumentations
scikit-learn==1.5.1
    # via albumentations
scipy==1.13.1
    # via
    #   albumentations
    #   scikit-image
    #   scikit-learn
sympy==1.13.1
    # via torch
threadpoolctl==3.5.0
    # via scikit-learn
tifffile==2024.7.24
    # via scikit-image
tomli==2.0.1
    # via albucore
torch==2.4.0
    # via torchvision
torchvision==0.19.0
tqdm==4.66.4
triton==3.0.0
    # via torch
typing-extensions==4.12.2
    # via
    #   albucore
    #   albumentations
    #   pydantic
    #   pydantic-core
    #   torch
